+++ src/main/java/peperact/EventHandler.java
@@ -8,12 +8,9 @@ import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import peperact.common.block.BlockBase;
-import peperact.common.block.peperact.BlockPeperact;
 import peperact.common.registry.RegistryBlocks;
-import peperact.common.registry.RegistryItems;
 
 import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
 
 @Mod.EventBusSubscriber(modid = Peperact.MODID)
 public class EventHandler {
@@ -38,7 +35,7 @@ public class EventHandler {
     }
 
     @SubscribeEvent
-    public static void onRegisterModel(ModelRegistryEvent event) {
+    public static void onRegisterModels(ModelRegistryEvent event) {
         Peperact.proxy.registerItemModels();
     }
 }