+++ src/main/java/peperact/client/ClientProxy.java
@@ -1,9 +1,7 @@
 package peperact.client;
 
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
-import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.client.model.ModelLoader;
-import peperact.Peperact;
 import peperact.common.CommonProxy;
 import peperact.common.registry.RegistryBlocks;
 import peperact.common.registry.RegistryItems;