+++ src/main/java/peperact/common/block/peperact/BlockPeperact.java
@@ -1,6 +1,5 @@
 package peperact.common.block.peperact;
 
-import net.minecraft.block.Block;
 import net.minecraft.block.ITileEntityProvider;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;