+++ src/main/java/peperact/common/block/peperact/TilePeperact.java
@@ -4,12 +4,9 @@ import com.google.common.base.Predicates;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.Tuple;
-import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.capabilities.CapabilityInject;
 import net.minecraftforge.energy.IEnergyStorage;
-import net.minecraftforge.fluids.Fluid;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import net.minecraftforge.fluids.capability.IFluidTankProperties;
@@ -18,7 +15,10 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -66,19 +66,15 @@ public class TilePeperact extends TileEntity implements IItemHandler, IFluidHand
     @Nullable
     @Override
     public <T> T getCapability(Capability<T> capability, @Nullable EnumFacing facing) {
-        if (capability == ITEM_HANDLER_CAPABILITY) return (T) this;
-        if (capability == FLUID_HANDLER_CAPABILITY) return (T) this;
-        if (capability == ENERGY_STORAGE_CAPABILITY) return (T) this;
+        if(capability == ITEM_HANDLER_CAPABILITY) return (T) this;
+        if(capability == FLUID_HANDLER_CAPABILITY) return (T) this;
+        if(capability == ENERGY_STORAGE_CAPABILITY) return (T) this;
         return super.getCapability(capability, facing);
     }
 
     @SuppressWarnings("unchecked")
     private <T> Stream<T> getTargetCapabilities(Capability<T> capability) {
-        return
-                //this.getWorld().loadedTileEntityList.stream()
-                //.filter(tile -> tile instanceof TilePeperact)
-                //.map(tile -> (TilePeperact) tile)
-                allTiles.stream()
+        return allTiles.stream()
                 .filter(p -> !p.inserting && !p.isInvalid())
                 .flatMap(p -> Arrays.stream(EnumFacing.VALUES)
                         .map(side -> Pair.of(side.getOpposite(), p.getPos().offset(side)))
@@ -88,32 +84,11 @@ public class TilePeperact extends TileEntity implements IItemHandler, IFluidHand
                         .map(sidetile -> sidetile.getRight().getCapability(capability, sidetile.getLeft()))
                         .filter(Predicates.notNull())
                 );
-//        ArrayList<T> caps = new ArrayList<>();
-//        Iterator<TilePeperact> iter = allTiles.iterator();
-//        while(iter.hasNext()) {
-//            TilePeperact peperact = iter.next();
-//            if (peperact.isInvalid()) {
-//                //iter.remove(); // Can't do this
-//                continue;
-//            }
-//            if (peperact.inserting) continue;
-//            // Get adjacent capabilities
-//            for(EnumFacing side : EnumFacing.VALUES) {
-//                BlockPos target = peperact.getPos().offset(side);
-//                if(!peperact.getWorld().isBlockLoaded(target)) continue;
-//                TileEntity tile = peperact.getWorld().getTileEntity(target);
-//                if(tile == null) continue;
-//                T cap = tile.getCapability(capability, side.getOpposite());
-//                if(cap == null) continue;
-//                caps.add(cap);
-//            }
-//        }
-//        return caps;
     }
 
     @Override
     public int receiveEnergy(int maxReceive, boolean simulate) {
-        if (this.getWorld().isRemote || inserting || maxReceive <= 0) return 0;
+        if(this.getWorld().isRemote || inserting || maxReceive <= 0) return 0;
         this.inserting = true;
 
         List<IEnergyStorage> caps = getTargetCapabilities(ENERGY_STORAGE_CAPABILITY)
@@ -123,7 +98,7 @@ public class TilePeperact extends TileEntity implements IItemHandler, IFluidHand
         int remaining = maxReceive;
         for(IEnergyStorage cap : caps) {
             remaining -= cap.receiveEnergy(remaining, simulate);
-            if (remaining <= 0) break;
+            if(remaining <= 0) break;
         }
 
         this.inserting = false;
@@ -132,7 +107,7 @@ public class TilePeperact extends TileEntity implements IItemHandler, IFluidHand
 
     @Override
     public int fill(FluidStack resource, boolean doFill) {
-        if (this.getWorld().isRemote || inserting || resource == null || resource.amount <= 0) return 0;
+        if(this.getWorld().isRemote || inserting || resource == null || resource.amount <= 0) return 0;
         this.inserting = true;
 
         List<IFluidHandler> caps = getTargetCapabilities(FLUID_HANDLER_CAPABILITY)
@@ -147,7 +122,7 @@ public class TilePeperact extends TileEntity implements IItemHandler, IFluidHand
         FluidStack remaining = resource.copy();
         for(IFluidHandler cap : caps) {
             remaining.amount -= cap.fill(remaining, doFill);
-            if (remaining.amount <= 0) break;
+            if(remaining.amount <= 0) break;
         }
 
         this.inserting = false;
@@ -157,7 +132,7 @@ public class TilePeperact extends TileEntity implements IItemHandler, IFluidHand
     @Nonnull
     @Override
     public ItemStack insertItem(int slotAlwaysZero, @Nonnull ItemStack stack, boolean simulate) {
-        if (this.getWorld().isRemote || inserting || stack.isEmpty() || stack.getCount() <= 0) return stack;
+        if(this.getWorld().isRemote || inserting || stack.isEmpty() || stack.getCount() <= 0) return stack;
         this.inserting = true;
 
         List<IItemHandler> caps = getTargetCapabilities(ITEM_HANDLER_CAPABILITY)
@@ -169,13 +144,12 @@ public class TilePeperact extends TileEntity implements IItemHandler, IFluidHand
             int numSlots = cap.getSlots();
             for(int slot = 0; slot < numSlots; slot++) {
                 remaining = cap.insertItem(slot, remaining, simulate);
-                if (remaining.getCount() <= 0) break;
+                if(remaining.getCount() <= 0) break;
             }
-
         }
 
         this.inserting = false;
-        return remaining;
+        return remaining.isEmpty() ? ItemStack.EMPTY : remaining;
     }
 
     @Override