+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next
+- Fixed block break particles
+
 ## v0.0.1 (2018-08-02)
 - Initial Release
 - Peperacts are all connected to each other for the moment